# Optimize option value detection