@byroot 2024-10-17
Superseeded by: https://github.com/ruby/json/pull/622

Thank you.